+++ src/serialchemy/_tests/test_serialization.py
@@ -1,5 +1,3 @@
-import datetime
-
 import pytest
 
 from serialchemy._tests.sample_model import Address, Company, Department, Employee
@@ -43,6 +41,16 @@ class EmployeeSerializerMixedFields(ModelSerializer):
     department = PrimaryKeyField(Department)
 
 
+class EmployeeSerializerHybridProperty(ModelSerializer):
+
+    full_name = Field(dump_only=True)
+
+
+class EmployeeSerializerProtectedField(ModelSerializer):
+
+    _role = Field()
+
+
 class CompanySerializer(ModelSerializer):
 
     employees = PrimaryKeyField(Employee)
@@ -51,7 +59,7 @@ class CompanySerializer(ModelSerializer):
 @pytest.fixture(autouse=True)
 def seed_data(db_session):
     company = Company(id=5, name='Terrans', location='Korhal')
-    emp1 = Employee(id=1, firstname='Jim', lastname='Raynor', company=company)
+    emp1 = Employee(id=1, firstname='Jim', lastname='Raynor', _salary=400, _role='Senior', company=company)
     emp2 = Employee(id=2, firstname='Sarah', lastname='Kerrigan', company=company)
     emp3 = Employee(id=3, firstname='Tychus', lastname='Findlay')
 
@@ -145,3 +153,22 @@ def test_empty_nested(db_session):
     assert serialized['company'] is None
     model = serializer.load(serialized, session=db_session)
     assert model.company is None
+
+def test_property_serialization(db_session):
+    serializer = EmployeeSerializerHybridProperty(Employee)
+    serialized = serializer.dump(db_session.query(Employee).get(2))
+    assert serialized['full_name'] is not None
+
+
+def test_protected_field_default_creation(db_session):
+
+    serializer = EmployeeSerializerProtectedField(Employee)
+    employee = db_session.query(Employee).get(1)
+    assert employee._salary == 400
+    serialized = serializer.dump(employee)
+    assert serialized.get('_salary') is None
+    assert serialized.get('_role') == 'Senior'
+
+    model = serializer.load(serialized, session=db_session)
+    assert model._salary is None
+    assert model._role == 'Senior'