+++ src/serialchemy/model_serializer.py
@@ -23,14 +23,10 @@ class ModelSerializer(Serializer):
         self._mapper_class = model_class
         self._fields = self._get_declared_fields()
         # Collect columns not declared in the serializer
-        for column_name, column in self.model_columns.items():
-            field = self._fields.setdefault(column_name, Field())
-            if field.serializer is None:
-                # If no serializer is defined, check if the column type has some serialized
-                # registered in EXTRA_SERIALIZERS.
-                for serializer_class, serializer_check in self.EXTRA_SERIALIZERS:
-                    if serializer_check(column):
-                        field._serializer = serializer_class(column)
+        for column_name in self.model_columns.keys():
+            if column_name.startswith('_'):
+                continue
+            self._fields.setdefault(column_name, Field())
 
     @property
     def model_class(self):
@@ -56,8 +52,9 @@ class ModelSerializer(Serializer):
         for attr, field in self._fields.items():
             if field.load_only:
                 continue
-            value = getattr(model, attr) if hasattr(model, attr) else None
+            value = getattr(model, attr, None)
             if field:
+                self._assign_default_field_serializer(field, attr)
                 serialized = field.dump(value)
             else:
                 serialized = value
@@ -86,6 +83,7 @@ class ModelSerializer(Serializer):
             field = self._fields[field_name]
             if field.dump_only:
                 continue
+            self._assign_default_field_serializer(field, field_name)
             if isinstance(field, SessionBasedField):
                 deserialized = field.load(value, session=session)
             else:
@@ -109,6 +107,21 @@ class ModelSerializer(Serializer):
         """
         return self.model_class()
 
+    def _assign_default_field_serializer(self, field, column_name):
+        """
+        If no serializer is defined, check if the column type has some serialized
+        registered in EXTRA_SERIALIZERS.
+
+        :param Field field: the field to assign default serializer
+
+        :param str column_name: sqlalchemy column name on model
+        """
+        column = self.model_columns.get(column_name)
+        if field.serializer is None and column is not None:
+            for serializer_class, serializer_check in self.EXTRA_SERIALIZERS:
+                if serializer_check(column):
+                    field._serializer = serializer_class(column)
+
     @classmethod
     def _get_declared_fields(cls) -> dict:
         fields = {}
@@ -116,8 +129,6 @@ class ModelSerializer(Serializer):
         if cls is ModelSerializer:
             return fields
         for attr_name in dir(cls):
-            if attr_name.startswith('_'):
-                continue
             value = getattr(cls, attr_name)
             if isinstance(value, Field):
                 fields[attr_name] = value