+++ CHANGELOG.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.2.0 (2019-03-22)
+------------------
+
+* Fix: Error when deserializing of nested models when SQLAlchemy model primary
+  key attribute name differs from the column name
+* Allow EXTRA_SERIALIZERS to be defined in runtime
+* Check if a session was given when serializing/deserializing nested fields
+
 0.1.0 (2019-02-12)
 ------------------